input and tag stories and tests
