+++ stories/Tag.jsx
@@ -6,16 +6,7 @@ import { noop } from '../test/util';
 import { item } from '../test/mock';
 
 storiesOf('Tag', module)
-  .add('focus and select text by clicking to enter edit mode', () => {
-    return (
-      <Tag
-        value={item}
-        handleEdit={noop}
-        handleRemove={noop}
-      />
-    )
-  })
-  .add('edit item by clicking and typing', () => {
+  .add('edit item by clicking text', () => {
     return (
       <Tag
         value={item}
@@ -30,36 +21,6 @@ storiesOf('Tag', module)
         value={item}
         handleEdit={noop}
         handleRemove={action('handleRemove')}
-      />
-    )
-  })
-  .add('exit edit mode by focus leaving textarea', () => {
-    return (
-      <Tag
-        value={item}
-        handleEdit={noop}
-        handleRemove={noop}
-        handleDoneEditing={action('handleDoneEditing')}
-      />
-    )
-  })
-  .add('exit edit mode by pressing enter', () => {
-    return (
-      <Tag
-        value={item}
-        handleEdit={noop}
-        handleRemove={noop}
-        handleDoneEditing={action('handleDoneEditing')}
-      />
-    )
-  })
-  .add('exit edit mode by pressing tab', () => {
-    return (
-      <Tag
-        value={item}
-        handleEdit={noop}
-        handleRemove={noop}
-        handleDoneEditing={action('handleDoneEditing')}
       />
     )
   });
